requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.3.1
matplotlib>=3.4.3
seaborn>=0.11.2
scikit-image>=0.18.3
pandas>=1.3.3
pytest>=6.2.5
tensorboard>=2.7.0
tqdm>=4.62.3
pyyaml>=5.4.1
opencv-python>=4.5.3